moved include files for cel parts to be compatible with the cel naming scheme to ease future backports merged back recent extensions and fixes to the recast and celgraph plugins made in cel (e.g. better handling of sector boundaries) reverted parallalized BuildTile as it didn't work properly

+++ src/plugins/common/celgraph/celgraph.cpp
@@ -245,7 +245,7 @@ iSector* celPath:: CurrentSector ()
 
 bool celPath::HasNext ()
 {
-  return cur_node < (int)(nodes.GetSize()-1);
+  return cur_node < nodes.GetSize()-1;
 }
 
 bool celPath::HasPrevious ()
